Refactor UnitTesting33 creation
This commit...

1. directly stores python 3.3 compat module's code in a string since
   sublime.load_resource() fails when called from plugin_loaded() in
   case reloading is triggered by removing UnitTesting from
   "ignored_packages".

   We could also delay execution, but the simple and tiny module seems
   better placed in a string to avoid the API call during startup.

2. creates files only, if they don't already exist to avoid unnecessary
   reloading UnitTesting33/ut.py module during startup.

3. Catch OSErrors during file creation to avoid tracebacks.
